requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.8
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
